+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/FilterMaker.java
@@ -29,32 +29,4 @@ public interface FilterMaker {
      * @throws FilterException 过滤器异常。
      */
     Filter makeFilter(FilterInfo filterInfo) throws FilterException;
-
-    /**
-     * 提供类型。
-     *
-     * @return 类型。
-     */
-    String provideType();
-
-    /**
-     * 提供标签。
-     *
-     * @return 标签。
-     */
-    String provideLabel();
-
-    /**
-     * 提供描述。
-     *
-     * @return 描述。
-     */
-    String provideDescription();
-
-    /**
-     * 提供示例内容。
-     *
-     * @return 示例内容。
-     */
-    String provideExampleContent();
 }