+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/filter/RangedDoubleFilterMaker.java
@@ -53,31 +53,6 @@ public class RangedDoubleFilterMaker implements FilterMaker {
         }
     }
 
-    @Override
-    public String provideType() {
-        return SUPPORT_TYPE;
-    }
-
-    @Override
-    public String provideLabel() {
-        return "具有范围的双精度浮点过滤器";
-    }
-
-    @Override
-    public String provideDescription() {
-        return "如果数据值是双精度浮点数且数值在配置的范围之内，则通过过滤。";
-    }
-
-    @Override
-    public String provideExampleContent() {
-        return JSON.toJSONString(new Config(
-                0.5,
-                true,
-                -1.25,
-                false
-        ), true);
-    }
-
     @Component
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public static class RangedDoubleFilter implements Filter, Bean {