+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/trigger/RangedLongTriggerMaker.java
@@ -53,31 +53,6 @@ public class RangedLongTriggerMaker implements TriggerMaker {
         }
     }
 
-    @Override
-    public String provideType() {
-        return SUPPORT_TYPE;
-    }
-
-    @Override
-    public String provideLabel() {
-        return "具有范围的长整型触发器";
-    }
-
-    @Override
-    public String provideDescription() {
-        return "如果数据值是长整型数且数值在配置的范围之内，则进行触发。";
-    }
-
-    @Override
-    public String provideExampleContent() {
-        return JSON.toJSONString(new Config(
-                1L,
-                true,
-                -2L,
-                false
-        ), true);
-    }
-
     @Component
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public static class RangedLongTrigger implements Trigger, Bean {